+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+- Fix a bug that would not match the content hash when another message was present
+
 ## 1.1.1 - 2022-01-18
 - Fix a bug that prevented uploading directories specified with a relative path